requirements.txt
google-generativeai>=0.8.6
pydub==0.25.1
tqdm==4.67.1